+++ internal/cli/util/view.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"github.com/ariel-frischer/autospec/internal/cli/shared"
-	"github.com/ariel-frischer/autospec/internal/config"
 	clierrors "github.com/ariel-frischer/autospec/internal/errors"
 	"github.com/ariel-frischer/autospec/internal/validation"
 	"github.com/spf13/cobra"
@@ -54,7 +53,7 @@ func init() {
 func runView(cmd *cobra.Command, args []string) error {
 	configPath, _ := cmd.Flags().GetString("config")
 
-	cfg, err := config.Load(configPath)
+	cfg, err := shared.LoadConfig(cmd, configPath)
 	if err != nil {
 		cliErr := clierrors.ConfigParseError(configPath, err)
 		clierrors.PrintError(cliErr)